Tests/LibWeb/Text/expected/SVG/use-element-shadow-tree-removal.txt

Initial shadow children: 1
After removal shadow children: 0
PASS: No stack overflow from deep recursion
